requirements.txt
discord.py
yt-dlp
spotipy
python-dotenv
PyNaCl
google-generativeai
Flask
waitress
requests
PyMuPDF